Criado middleware para controle de acesso admin

+++ routes/api.php
@@ -39,22 +39,22 @@ Route::group([
 //brand routes
 Route::group(['prefix' => 'marcas'], function ($routes) {
     Route::controller(BrandController::class)->group(function () {
-        Route::get('index', 'index');
-        Route::get('show/{id}', 'show');
-        Route::post('store', 'store');
-        Route::put('update_brand/{id}', 'update_brand');
-        Route::delete('delete_brand/{id}', 'delete_brand');
+        Route::get('index', 'index')->middleware('is_admin');
+        Route::get('show/{id}', 'show')->middleware('is_admin');
+        Route::post('store', 'store')->middleware('is_admin');
+        Route::put('update_brand/{id}', 'update_brand')->middleware('is_admin');
+        Route::delete('delete_brand/{id}', 'delete_brand')->middleware('is_admin');
     });
 });
 
 //categories routes
 Route::group(['prefix' => 'categorias'], function ($routes) {
     Route::controller(CategoryController::class)->group(function () {
-        Route::get('index', 'index');
-        Route::get('show/{id}', 'show');
-        Route::post('store', 'store');
-        Route::put('update_category/{id}', 'update_category');
-        Route::delete('delete_category/{id}', 'delete_category');
+        Route::get('index', 'index')->middleware('is_admin');
+        Route::get('show/{id}', 'show')->middleware('is_admin');
+        Route::post('store', 'store')->middleware('is_admin');
+        Route::put('update_category/{id}', 'update_category')->middleware('is_admin');
+        Route::delete('delete_category/{id}', 'delete_category')->middleware('is_admin');
     });
 });
 
@@ -72,9 +72,9 @@ Route::group(['prefix' => 'produtos'], function ($routes) {
     Route::controller(ProdutcController::class)->group(function () {
         Route::get('index', 'index');
         Route::get('show/{id}', 'show');
-        Route::post('store', 'store');
-        Route::put('update_product/{id}', 'update_product');
-        Route::delete('delete_product/{id}', 'delete_product');
+        Route::post('store', 'store')->middleware('is_admin');
+        Route::put('update_product/{id}', 'update_product')->middleware('is_admin');
+        Route::delete('delete_product/{id}', 'delete_product')->middleware('is_admin');
     });
 });
 
@@ -85,8 +85,8 @@ Route::group(['prefix' => 'compras'], function ($routes) {
         Route::get('index', 'index');
         Route::get('show/{id}', 'show');
         Route::post('store', 'store');
-        Route::get('get_order_items/{id}', 'get_order_items');
-        Route::get('get_user_orders/{id}', 'get_user_orders');
-        Route::post('change_order_status/{id}', 'change_order_status');
+        Route::get('get_order_items/{id}', 'get_order_items')->middleware('is_admin');
+        Route::get('get_user_orders/{id}', 'get_user_orders')->middleware('is_admin');
+        Route::post('change_order_status/{id}', 'change_order_status')->middleware('is_admin');
     });
 });
